Backbone work for User Profiles, Profile Pictures, and League Members display
Visiting /profile/[user_id] will direct to a user's profile page. Currently the profile page is displaying 'dummy' data except the profile picture (more info below).

- Profile Picture -
The profile picture component will return rounded profile image in the size requested. It is currently fetching the image from the public folder. The image needs to be saved as [user_id].jpeg to be found. If there is not profile picture, it will resolve to the "default_profile_picture.jpeg" photo.

/League and /profile both currently lead to the same table display. I wasn't sure what location would be best for this table at the moment. Quite possible it is neither of these lol. Can reevaluate later. They both utilize the new LeagueMembersTable component.